Proposed change to the point of view of 3 principles
There are 3 principles, within the "Minimise blockers and dependencies"
section, that are written from something more like an "outside" perspective.
The rest of the document is worded as something *we* have written and
that *we* stand by.

The 3 principles I've edited here read more like a rule handed down from
somewhere outwith of engineering.

I propose we word them so they read as if we own them.

I've also added a shift away from a wording that was akin to "they don't do
this, but they should", to "we act to embody this".